requirements.txt
astroid==1.6.3
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.3.0
geoip2==2.8.0
gunicorn==19.7.1
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
lazy-object-proxy==1.3.1
maxminddb==1.3.0
mccabe==0.6.1
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.4
python-decouple==3.1
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
ua-parser==0.8.0
urllib3==1.22
user-agents==1.1.0
wcwidth==0.1.7
wrapt==1.10.11
